rename adapter package
