+++ internal/helm/helmadapter/gorm_repo_store.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helmrepoadapter
+package helmadapter
 
 import (
 	"context"
